Reorganizing where types are stored.
